requirements.txt
requests
beautifulsoup4
lxml
python-dotenv
APScheduler -BlockingScheduler
